Improvements to OptickleTest:
Move printing and path checking error to constructor.
Change order of numerical verifyEqual, to make diagnostics more correctly labeled.
Optickle is not usually supposed to be in the path, both ref and calc paths are set in config.
error tolerances also in config.
diagnostic temporary file name stored to global so scripts can access it.